Remove microshift hosts from sf group
There is an issue with podman 5.1.0 on centos-9-stream for it's not
compatible with cni anymore:

[cloud-user@microshift ~]$ podman ps
Error: cni support is not enabled in this build, only netavark. Got
unsupported network backend "cni"

The cni plugin is hardcoded on [1], this change remove the hosts from
sf group to avoid node-exporter setup until the situation will be fixed.

[1] https://github.com/openstack-k8s-operators/ansible-microshift-role/blob/master/files/crio-bridge.conf

Change-Id: I2bd08a59589e138f0e420b4fecd8123dd6da3d14