Reboot CRC host before executing tests when extracted image
The kubelet should not be started before applying the changes in the
dnsmasq or NetworkManager. Normally, the crc-pre.service [1] [2] is doing
that, but it makes some complication on some infras to use that script,
so we disable it [3] and it seems to be very important to run the whole
commands what are in the script.

[1] https://github.com/softwarefactory-project/sf-infra/blob/master/roles/next-gen/extracted-crc/files/crc-pre.service
[2] https://github.com/softwarefactory-project/sf-infra/blob/master/roles/next-gen/extracted-crc/files/configure-pre-crc.sh
[3] https://softwarefactory-project.io/r/c/software-factory/sf-infra/+/31954

Change-Id: I4be2c406a79e7621fb25ff4a8afa79962c6be9b1