Disable OpenShift console; change authorized key file to zuul in crc
The ignition file might be overwritten by the machineconfigpool, so
later we might not get access to that host.
By moving the authorized keys to another file, we should probably avoid
that situation.

NOTE: The zuul authorized keys file does not have proper SELinux label
and we can not change it by mounting the home storage to the system.
The proper labels will be done by using our cloudinit ssh service.

Also disable the CRC console access. It is not neede by the Next Gen
team.

Change-Id: I2c2ed57fa6f37e974851722b274dc719998b14cd