Add ip address for documentation server
Documentation has been moved to github pages so www.rdoproject.org is
now a CNAME to the github infra.

We still need to maintain this server until we move the useful
dashboards in dashboards.rdoproject.org which run in the same server.
Until we can migrate and remove the server, let's manage the ip address
manually.

Change-Id: I575b9752aeee51830668ca8e94ff740306161f55