Add checking dashboards.rdoproject.org by the blackbox exporter
The www.rdoproject.org site has been moved to the Github mkdocs.
It means, that the previous DNS record has new ip address, so it
is impossible to monitor the host state by using node exporter.
On the www.rdoproject.org there is another site (dashboards.rdoproject.org),
that is still important and we can start monitor it.

Change-Id: Ie0f4dbfa558c66528428fc706d0c34006dbf0501