Change www.rdoproject.org to dashboards.rdoproject.org
This commit is a continuation of proposed patch [1] - we need to change
the host name to the domain on which the old www.rdoproject.org is set.
In other words, we still want to monitor www.rdoproject.org host, but
the DNS A recort is set to the Github, so the blackbox monitoring
service is failing on veryfing the domain.

[1] https://softwarefactory-project.io/r/c/software-factory/sf-infra/+/32140

Change-Id: I86461d92b946e9a259e751258afae2ae838d63f6