k9s: download pre-built binary instead of building it
There's no reason to build k9s ourselves when the project provides
a prebuilt binary.
Install the binary in /usr/local/bin instead of just inside the ansible user's
home.
Delete previous install to avoid conflict.

The only drawback is that github doesn't provide a link to a "latest"
release, so we will have to maintain the latest version ourselves in
the role's definition.

Change-Id: I48b76363a96ff778501150eef273159c148a4bc1