Enable configure dnsmasq for crc extracted image
The image needs to include enabling service that is responsible to
rewrite the dnsmasq configuration to host ip address.
That is because when the VM is starting and the nodepool is waiting
for other VMs to start, the CRC is making a lot request to resolve
api-int.crc.testing domain, which is normally rejected by the DNS
from our cloud provider. To imagine the scale of the issue, there is
over 20 milion requests done in 1 hour.
Similar change is also done in the CI job [1], but it is done too late.

[1] https://github.com/openstack-k8s-operators/ci-framework/blob/57ef3f1382d0f4d6d3a25e514c6baa18fa092e06/roles/reproducer/tasks/configure_crc.yml#L31

Change-Id: Iebab70aae519844b8b219bc8eb59b3969848c88a