Add script for checking old containers on k1s host; add monitor probe
Sometimes there are some containers leftovers due Zuul restart
or connection issues. Those containers are running and they are counted
to the max-servers limit and also they are taking resources.
This commit adds script that is checking if there some containers older
than 1 day and if yes, it will ring an alert.

Change-Id: Ia897bd6a17b1650b4c764b8c051309a839d845eb