Trigger cleanup ports more often; get port stats more often
We are out of the ip adresses from time to time on some infras.
The new CI jobs for the CI framework team are creating ports and
attaching to the instance, so Nova is not an owner of it, so it might
be some leftovers, when there would be some communication problem with
the Zuul to trigger the post job (which is removing the created
resources like routers, ports, trunk ports etc.).

Change-Id: Id0578a9cd31f8bfb11dd09dba32de0c5cb27ed1f