Add feature for adding cloud_image_builder_users user into sudoers
The crc-cloud images require to execute some tasks [1][2] as privileged
user, due even after mounting the qcow2 image to the system,
uid and gid are still poiting to user that is available in the system.
That means, that if we want to create a files in /etc/systemd dir,
we need to have permissions to do that.

There can be done ugly workaround like create a user systemd service
(configure dnsmasq or enable cloud init ssh keys if someone is using
config drive), which will call sudo to execute proper commands, when the
system is starting. By moving all scripts to new service will
make another abstraction layer that will be more difficult for
debugging for other users.

[1] https://softwarefactory-project.io/r/plugins/gitiles/config/+/refs/heads/master/nodepool/virt_images/roles/extracted-crc/tasks/service.yaml#3
[2] https://softwarefactory-project.io/r/plugins/gitiles/config/+/refs/heads/master/nodepool/virt_images/roles/extracted-crc/tasks/modify.yaml#15

Change-Id: I487321539e8e5fa7c98bb058cdb4c22c51a92816