Replace puppeteer script with grafana-image-rendering
The puppeteer script was difficult to maintain. Current issue was that
the puppeteer was making empty screenshots because there is no datasource
available. That issue was caused by the PR [1].
The host_vars is already generated by dhall and it is difficult to add
there vault content where normally it is its place. That's why the vault
content was moved to group_vars file some time ago.
This commit move the vault content to group_vars all file - the variables
are not duplicating.

[1] https://softwarefactory-project.io/r/c/software-factory/sf-infra/+/32493

Change-Id: I084c2cd6b9362973cfe2ce0df64c0d199bfcc003