Increase retention time to 32 hours
The script in some cases might remove ports that are attached
to the VM that are in hold state, so the developers can not
continue investigation.
To check what ports and networka are in use, we need to make more
queries, some are going to Zuul CI, which in some infras will require
to implement GSSAPI, then make queries to nodepool service to get
VMs uuids and filter them out in cleaner to not being "touched" by it.

That patch set is a workaround to avoid removing hold nodes ports.

Change-Id: I81c749315e7a52aacc6e58abf3b81840df78f22a