Remove sf-infra-configure-quay job and migrate tasks in site_bridge
The crontab to be installed on the bridge for the quay.rdoproject.org
certificate renewal was installed by this job. Last time the job was
running was 1 year ago.

The bridge have been reinstalled so the crontab have not been
installed.

This change migrates the tasks where they will be run often.

Change-Id: Id2e16c6fdaa191f4248d0b25c3f15f132d9d53ed