Temporary ignore CA cert verification for IBM cloud; add sf-operator.yaml match
The sf-operator does not provide functionality to provide custom
cacert files that sometimes is required by the openstack client
to connect to the cloud provider.

Also add matching the group_vars/sf-operator.yaml file to trigger the CI
job.

Revert that change or replace it with 'cacart' when such functionality
is available in sf-operator.

Change-Id: I20b917ee0f2a99122cb55be322b51259c69f3707