centosinfra-prod: add gitlab connection
Just like with the pagure connections, we don't control the webhook definitions from our side. We'll ask stakeholders to change their config
when their tenant is migrated.

The good thing is that gitlab supports multiple webhooks, so we can keep both the old and the new one running in parallel.

Change-Id: Ibe8b643f3a04e56b6ccb2c056a4b690474225368