openafs: Ensure module installation until upstream bug is fixed
There is an issue with dkms on rhel-9.4 and centos-9-stream on
dkms.conf file provided by openafs.

The option NO_WEAK_MODULES is set to true and now dkms accept only
'yes' or 'no' value [1]

I created a bug report upstream for openafs [2] before finding the fix, I'm waiting
for the triage to add the solution to fix the issue.

This change ensure the module can be installed as expected

[1] https://github.com/dell/dkms/blob/main/dkms.in#L824
[2] https://rt.central.org/rt/SelfService/Display.html?id=135545

Change-Id: Id86a4db16b30038dca40e8a6f1d49684f260b445