secret-age: remove the blame technic and expect a date
The change changes the behavior to not fallback to the date computed
by git blame. The metric is set to 0 when the date marker is not
found.

This could enable the configuration of an alert to raise when no
marker are set on secrets (0 value) as well as the regular alerts when
the secret require a rotation or an action.

Also:

- add additional test for the refreshed_date marker

Change-Id: I9104e3d07037c11dee4693b45525ff3f82f30c81