Import Zookeeper migration tools from upstream
This change imports two tools to help decrypt and encrypt Zookeeper
Keys.

This is useful when we what to migrate Zuul projet to a new instance.

We made these change upstream that are still not merged, so we decided,
to have them closer, and start to be manage by us.

Check changes at:
https://review.opendev.org/q/topic:%22zuul-decrypt-secret%22

Change-Id: I1440478d176b48d18de2cc3289c391e5b17be2bd