Clarify k1s secret rotation procedure
The recent k1s certificate rotation caused an outage due to an SSL
mismatch. This happened because the private key (`k1s_key`) was not
updated along with the certificate and token.

To prevent this from happening again, add a prominent note to the
documentation. The note clarifies that all `k1s` values must be
rotated together, even if their "refreshed" dates suggest otherwise.

Change-Id: I44b238957d510be70f4a2c78c692607f777ea587